Updated AWS test services according to interface changes from camel 3.7
