+++ tests/itests-aws-v1/src/test/java/org/apache/camel/kafkaconnector/aws/v1/s3/source/CamelSourceAWSS3ITCase.java
@@ -18,16 +18,22 @@
 package org.apache.camel.kafkaconnector.aws.v1.s3.source;
 
 import java.io.File;
+import java.util.Iterator;
 import java.util.Properties;
 import java.util.concurrent.ExecutionException;
 
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.s3.AmazonS3;
-import org.apache.camel.kafkaconnector.aws.v1.clients.AWSClientUtils;
+import com.amazonaws.services.s3.model.ListVersionsRequest;
+import com.amazonaws.services.s3.model.ObjectListing;
+import com.amazonaws.services.s3.model.S3ObjectSummary;
+import com.amazonaws.services.s3.model.S3VersionSummary;
+import com.amazonaws.services.s3.model.VersionListing;
 import org.apache.camel.kafkaconnector.common.AbstractKafkaTest;
 import org.apache.camel.kafkaconnector.common.ConnectorPropertyFactory;
 import org.apache.camel.kafkaconnector.common.clients.kafka.KafkaClient;
 import org.apache.camel.kafkaconnector.common.utils.TestUtils;
+import org.apache.camel.test.infra.aws.clients.AWSClientUtils;
 import org.apache.camel.test.infra.aws.common.AWSCommon;
 import org.apache.camel.test.infra.aws.common.AWSConfigs;
 import org.apache.camel.test.infra.aws.common.services.AWSService;
@@ -49,13 +55,66 @@ import static org.junit.jupiter.api.Assertions.fail;
 public class CamelSourceAWSS3ITCase extends AbstractKafkaTest {
 
     @RegisterExtension
-    public static AWSService<AmazonS3> service = AWSServiceFactory.createS3Service();
+    public static AWSService service = AWSServiceFactory.createS3Service();
     private static final Logger LOG = LoggerFactory.getLogger(CamelSourceAWSS3ITCase.class);
 
     private AmazonS3 awsS3Client;
 
     private volatile int received;
     private final int expect = 10;
+
+    /**
+     * Delete an S3 bucket using the provided client. Coming from AWS documentation:
+     * https://docs.aws.amazon.com/AmazonS3/latest/dev/delete-or-empty-bucket.html#delete-bucket-sdk-java
+     *
+     * @param s3Client
+     *            the AmazonS3 client instance used to delete the bucket
+     * @param bucketName
+     *            a String containing the bucket name
+     */
+    public static void deleteBucket(AmazonS3 s3Client, String bucketName) {
+        // Delete all objects from the bucket. This is sufficient
+        // for non versioned buckets. For versioned buckets, when you attempt to delete objects, Amazon S3 inserts
+        // delete markers for all objects, but doesn't delete the object versions.
+        // To delete objects from versioned buckets, delete all of the object versions before deleting
+        // the bucket (see below for an example).
+        ObjectListing objectListing = s3Client.listObjects(bucketName);
+        while (true) {
+            Iterator<S3ObjectSummary> objIter = objectListing.getObjectSummaries().iterator();
+            while (objIter.hasNext()) {
+                s3Client.deleteObject(bucketName, objIter.next().getKey());
+            }
+
+            // If the bucket contains many objects, the listObjects() call
+            // might not return all of the objects in the first listing. Check to
+            // see whether the listing was truncated. If so, retrieve the next page of objects
+            // and delete them.
+            if (objectListing.isTruncated()) {
+                objectListing = s3Client.listNextBatchOfObjects(objectListing);
+            } else {
+                break;
+            }
+        }
+
+        // Delete all object versions (required for versioned buckets).
+        VersionListing versionList = s3Client.listVersions(new ListVersionsRequest().withBucketName(bucketName));
+        while (true) {
+            Iterator<S3VersionSummary> versionIter = versionList.getVersionSummaries().iterator();
+            while (versionIter.hasNext()) {
+                S3VersionSummary vs = versionIter.next();
+                s3Client.deleteVersion(bucketName, vs.getKey(), vs.getVersionId());
+            }
+
+            if (versionList.isTruncated()) {
+                versionList = s3Client.listNextBatchOfVersions(versionList);
+            } else {
+                break;
+            }
+        }
+
+        // After all objects and object versions are deleted, delete the bucket.
+        s3Client.deleteBucket(bucketName);
+    }
 
     @Override
     protected String[] getConnectorsInTest() {
@@ -64,7 +123,7 @@ public class CamelSourceAWSS3ITCase extends AbstractKafkaTest {
 
     @BeforeEach
     public void setUp() {
-        awsS3Client = service.getClient();
+        awsS3Client = AWSClientUtils.newS3Client();
         received = 0;
 
         try {
@@ -78,7 +137,7 @@ public class CamelSourceAWSS3ITCase extends AbstractKafkaTest {
     @AfterEach
     public void tearDown() {
         try {
-            AWSClientUtils.deleteBucket(awsS3Client, AWSCommon.DEFAULT_S3_BUCKET);
+            deleteBucket(awsS3Client, AWSCommon.DEFAULT_S3_BUCKET);
         } catch (Exception e) {
             LOG.warn("Unable to delete bucked: {}", e.getMessage(), e);
         }
